testApiApp/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-environ==0.11.1
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
idna==3.4
inflection==0.3.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
